+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * feature: support PDF output with `website-evidence-reporter` using chromium (brings the benefit of adding page title and page numbers)
 * dependencies: update dependencies (`npm update`)
+* fix: fix `WORKDIR` in Dockerfile (fixes #106)
 
 ## 2.1.1 / 2023-09-20